Code review: Convert to regex when parsing the rule and pass moshi as a dependency to RealRequestBlocklist
